+++ src/Catalog/DataAbstractionLayer/CategoryProductMappingIndexer.php
@@ -5,7 +5,6 @@ namespace Shopgate\Shopware\Catalog\DataAbstractionLayer;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use JsonException;
-use Monolog\Level;
 use Shopgate\Shopware\Catalog\Category\ProductMapBridge;
 use Shopgate\Shopware\Shopgate\Catalog\CategoryProductIndexingMessage;
 use Shopgate\Shopware\Storefront\ContextManager;
@@ -103,11 +102,17 @@ class CategoryProductMappingIndexer extends EntityIndexer
             return;
         }
 
-
-        $delCount = Profiler::trace('shopgate:catalog:product:indexer:delete', function () use ($ids, $channels): int {
-            $this->logger->logBasics('Removing categories', ['categories' => array_values($ids)]);
-            return $this->productMapBridge->deleteCategories($ids, $channels);
-        });
+        // todo: add config?
+        $delCount = 0;
+        if ($message->isFullIndexing) {
+            $delCount = Profiler::trace(
+                'shopgate:catalog:product:indexer:delete',
+                function () use ($ids, $channels): int {
+                    $this->logger->logBasics('Removing categories', ['categories' => array_values($ids)]);
+                    return $this->productMapBridge->deleteCategories($ids, $channels);
+                }
+            );
+        }
 
         $writeCount = Profiler::trace(
             'shopgate:catalog:product:indexer:update',
@@ -162,11 +167,8 @@ class CategoryProductMappingIndexer extends EntityIndexer
                     $langSortOrder,
                     function (array $category) use ($catId, $channelId, $rawCat, $langId, $defaultLang) {
                         $sameSlot = $category['slot'] === $rawCat['slot_config'] && $category['slot'] === null;
-                        // for a specific channel, if default language sorting is set and non-default lang sort is null, skip map
-                        // because SW falls back to default language sorting
-//                        $nonMainIsNull = $defaultLang !== $langId && $rawCat['slot_config'] === null && $sameChannel;
                         $sameChannel = $channelId === $category['channelId'];
-                        return $category['catId'] === $catId && $sameSlot && $sameChannel/*|| $nonMainIsNull*/;
+                        return $category['catId'] === $catId && $sameSlot && $sameChannel/*|| $nonMainIsNull*/ ;
                     }
                 );
                 if (!empty($findNotUnique)) {